requirements.txt
fastapi
uvicorn
python-multipart
numpy
torch
transformers
pillow
